解决：错字，点击节点影响排序增加随机性，优化统计节点公司数的代码，撤销转化Num由string为数字（因为接口数据已修改） 待解决：详情的新闻信息新闻摘要展开和收拢功能不灵敏，需要改变显示框大小才可以。

+++ src/utils.js
@@ -68,9 +68,9 @@ function compare (propertyName, sort) {
         var value2 = obj2[propertyName]
         if (typeof value1 === 'string' && typeof value2 === 'string') {
             if(propertyName === 'capital'){
-                if (parseInt(value1) <= parseInt(value2)) {
+                if (parseFloat(value1) <= parseFloat(value2)) {
                     return sort === 'ascending' ? -1 : 1
-                } else if (parseInt(value1) > parseInt(value2)) {
+                } else if (parseFloat(value1) > parseFloat(value2)) {
                     return sort === 'ascending' ? 1 : -1
                 }
             }
@@ -85,7 +85,6 @@ function compare (propertyName, sort) {
         }
     }
 }
-
 export default {
     limitNum,
     compare
